backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# AI/ML Dependencies for Hugging Face
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# Cross-modal detection dependencies
Pillow>=9.0.0
numpy>=1.21.0
openai-whisper>=20231117 